backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
groq==0.9.0
python-jose==3.3.0